requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
psycopg[binary]